requirements.txt
Flask==2.3.3
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
numpy==1.26.4
Pillow==10.0.1
PyMySQL==1.1.0
Werkzeug==2.3.7